Implemented fractional coordinates Added a test target

+++ lib/functions.cpp
@@ -26,7 +26,7 @@ double dist(double x1, double x2){
 
 std::string split(std::string s, char delim){
     int p;
-    p = s.find_last_of(delim);
+    p = (int) s.find_last_of(delim);
     return s.substr(0,p);
 }
